Updated Angular Material (and CDK?) to v10.
For some reason the CDK updates with Material and not with core. Odd.

Also removed codelyzer from dependencies entirely - my understanding
is that it's a kind of linting tool that 1. I don't use and 2. has
been deprecated in favour of something like eslint and 3. causes
dependency issues as a consequence.

Also manually set jasmine-core back up to 3.8.0 again; I think the
previous update downgraded it, but I can't npm install with a lower
version.