Updated Medication-Add component to work with new typed forms (prototype)
The only actually-relevant change (aside from having to provide default
values for fields so they're correctly type-inferred) is how we fetch the
value stored in the form. The old method was to simply use form.value. Now
that returnes an explicitly-optional result in order to account for the
possibility of some form controls (fields) being disabled. The FIXME
comment in the code discusses options more in-length.