Removed [errorStateMatcher] element from input fields.
It was being declared in the old code, but did not match any actual
ErrorStateMatcher object in the components, since none was declared. I
suspect this was just copy-pasted boilerplate that didn't do anything.
Still, there is a non-zero possibility that this was meant to somehow
interact with the validators (e.g. for nonempty names) that somehow just
worked in the original Angular without specifying the matcher object.