Replaced nth-last-child() with last-child in CSS selectors
It was being used without any argument, for the purpose (as far as I can
tell) of adding bottom padding to just the very last element. Doing
something to specfically the very last element is what last-child is for.

On the strange off-chance that it was actually used to select ALL elements,
that could be adjusted by just selecting for all children instead (no need
to count them backwards).