Switched to using SCSS, (partially) fixed button icons being wrong size.
Switched to using Sass for styling, since this is the new default for
Angular / Angular Material, I understand. Created a theme file for storing
styling meant to be a global change to how certain Angular Material basic
objects look.

Changed the built-in button spacing settings so the icons would no longer
be pushed to the side. This might negatively affect some future buttons,
being global, but is fine for now.

Changed the styling of mat-icon objects to be sized to 24px, like they used
to be before migration. The issue is that *something* in the style stack
introduces a sizing to 1.125rem that overrides the default 24px sizing of
the icons. I couldn't figure out exactly where this happens - it doesn't
seem to be in the prebuilt theme (deeppurple-amber.css) and I can't find
it in any of my own styles, so I and Deepseek suspect it might be a default
from Angular / Angular Material.
As such, since I couldn't just disable it, I used a more specific selector
to override it, taking advantage of the fact that most of the icon buttons
in the app are contained within a "button-row" div. The only exception are
the Edit and Delete buttons in the detail views, which therefore still have
the new, smaller size. Visually this is alright, but it's inconsistent and
therefore I'm labelling the fix as only partial.