Fixed inputs being solid-background and without dynamic labels.
The solid background was fixed directly from the theme SCSS. Notably, it
only affects the inputs while they are NOT focused; I personally think this
works better, but it does not work well with patient-activities-detail,
where this makes it obvious that the text area stretches all the way up but
not all the way down in the table row (this looks bad). That component
needs extra attention anyway, though.

The labelling was fixed by moving from having Angular automatically treat
the placeholder value as a label (old functionality) to manually including
a `mat-label` element. In the process of doing this, I also added the
`hideRequiredMarker` flag to mandatory fields, to prevent them from being
displayed with an asterisk.