Updated patient-activities-detail with fixed table/charts.
Computed activity duration from activity start and end and formatted it
before displaying.
I attempted to use a get property directly in Activity for this, but it did
not work (always returned undefined). My guess is that the way the service
initialises the activities does not actually involve calling new() and thus
the objects don't actually get populated with methods - they just get
filled with the data.
I used a custome DatePipe format at first ('H\'h\' m\'m\'), but there was
no way to only display hours if the activity lasted at least 1 hour, so
I wrote a custom formatting method.

Updated conditionals for when the various Google charts should be
displayed: now nothing is displayed until the activities are loaded, then
each chart/table is only displayed if there are any activities in its
selected timeframe. The headings and time period selectors are always
displayed as long as some activity data for the patient exists.

Changed patient activities table from using `<mat-table>` to using
`<table mat-table>` in hopes that it would improve column auto-sizing. It
did.

Added `hideIcon` to the slide toggle for whether activities are justified
to hide the checkmark / dash that the new Angular Material version added
and revert to the old look.