Updated patient activities justification textarea so it sizes correctly.
Added `subscriptSizing="dynamic"` flag to have the form field not allocate
space for an error message until it's needed. (Since there are no
restrictions on this field, it will never be needed, so the space was
wasted.)

Added CSS rule to inherit font size to circumvent the issue of component
height being calculated incorrectly from the `cdkAutosizeMaxRows`
attribute. If the underlying Angular Material bug is fixed, this CSS may
be removed. See https://github.com/angular/components/issues/26233.

Added CSS to make the preferred width of the textarea larger (from 180px
default to 240px) for adding notes about whether an activity is justified
or not. This looks okay on my fullscreen setup, but may be less good on a
smaller screen. As the TODO says, maybe a different solution would be
better.