Updated Docker / Nginx config.
Changed Docker to build the app in production mode. This should have been
the default to begin with.

Changed the Nginx configuration to serve files from the `browser` folder,
since it seems that's where they go now with the current build process.
I simply used a shell to check the directory / file structure, saw the
correct files in the `browser` folder, and moved the config to serve from
there; it is a valid alternate solution, I think, to change what gets
copied / where it gets copied from the build container, so that the Angular
index.html ends up in the Nginx default root foolder.